LayoutTests/fast/js/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.25-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S7.5.2_A1.25

PASS Expected parsing failure

TEST COMPLETE
